+++ src/components/Navbar.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useState } from "react";
 import { Twirl as Hamburger } from "hamburger-react";
 import Button from "./Button";
-import ContactUs from "./ContactUs";
+import ContactUs from "./navbar_items/ContactUs";
 
 const Navbar = () => {
   const [isOpen, setOpen] = useState(false);
